Final Edit. Needs some cleanup

+++ src/modules/camera.py
@@ -36,6 +36,7 @@ def _config_logger(log_path, log_name):
 
     console_handler = logging.StreamHandler()
     root_logger.addHandler(console_handler)
+    return root_logger
 
 
 def _encode_to_server(file_path):
@@ -45,31 +46,38 @@ def _encode_to_server(file_path):
     return time_stamp, image_64
 
 
-def _handle_camera(class_id, camera, sampling_rate, delay, client):
+def _handle_camera(class_id, camera, sampling_rate, delay, client, logger):
     camera.start_preview()
+    file_path = None
     while True:  # todo handle error
-        # pir.wait_for_motion(sampling_rate)
-        file_path = f'/tmp/image_{int(time.time())}.jpg'
-        camera.capture(file_path)
-        time_stamp, encode_image = _encode_to_server(file_path)
-        camera_capture(class_id, time_stamp, encode_image, client)
-        os.remove(file_path)
-        time.sleep(delay)
+        try:
+            # pir.wait_for_motion(sampling_rate)
+            if file_path is not None:
+                os.remove(file_path)
+            file_path = f'/tmp/image_{int(time.time())}.jpg'
+            print(file_path)
+            print(camera.capture(file_path,))
+            time_stamp, encode_image = _encode_to_server(file_path)
+            camera_capture(class_id, time_stamp, encode_image, client, logger)
+            
+            time.sleep(delay)
+        except Exception as e:
+            logger.error(e.traceback())
     camera.stop_preview()
 
-def camera_capture(class_id, time_stamp, encode_image, client):
+def camera_capture(class_id, time_stamp, encode_image, client, logger):
         # headers = self.header()
         data = {'class_id': class_id,
                 'time_stamp': time_stamp,
                 'encode_image': encode_image
                 }
         # print(data)
-        logging.error("camera_capture requested ")
+        logger.info("camera_capture requested ")
         # response = requests.post(f'{self.base_url}{BaseApi.camera_endpoint}',
         #                          headers=headers, data=json.dumps(data, ensure_ascii=False).encode('utf-8'))
         response = client.publish("fumSmartClassIot/camera", json.dumps(data))
         # response = client.publish("fumSmartClassIot/camera", "HELOOOO")
-        logging.error("camera_capture response" + str(response))
+        logger.info("camera_capture response" + str(response))
         
 
 def start_camera(pir_pin, height, width, sampling_rate, delay):
@@ -78,15 +86,16 @@ def start_camera(pir_pin, height, width, sampling_rate, delay):
         level=getattr(logging, config['logger']['log_level'], 'DEBUG'),
         format="%(asctime)s [%(threadName)-12.12s] [%(levelname)-5.5s]  %(message)s"
     )
-    _config_logger(config['logger']['logPath'], config['logger']['fileName'])
+    logger = _config_logger(config['logger']['logPath'], config['logger']['fileName'])
     broker_address = "broker.mqtt-dashboard.com"
-    client = mqtt.Client("SmartClassFUMCameraModule")
+    # broker_address = "mqtt.eclipse.org"
+    client = mqtt.Client("SmartClassFUMCameraModule123")
     client.connect(broker_address)
     camera = PiCamera()
     logging.error("Starting Camera")
     # pir = MotionSensor(int(pir_pin))
     camera.resolution = (int(height), int(width))
-    _handle_camera(config['class']['class_id'], camera, int(sampling_rate), int(delay), client)
+    _handle_camera(config['class']['class_id'], camera, int(sampling_rate), int(delay), client, logger)
     # thread = Thread(target=_handle_camera, args=(api, camera, pir, int(sampling_rate), int(delay)))
     # thread.daemon = True
     # thread.start()
